generic/077: ignore errors occurred while accessing the filler files
When looking for data to fill in the filesystem, errors accessing files
may occur.  This will cause the test to fail as it'll show in the output
lines such as:

    du: cannot read directory '/usr/etc/sudoers.d': Permission denied

Ignoring these errors should be safe, so simply redirecting the stderr of
'du' to $seqres.full fixes it.  Unfortunately, this exposed a different
issue, which was the truncation of the $seqres.full file while copying files
into the filesystem.  This patch also fixes that.

Signed-off-by: "Luis Henriques (SUSE)" <luis.henriques@linux.dev>
Reviewed-by: Zorro Lang <zlang@redhat.com>
Signed-off-by: Zorro Lang <zlang@kernel.org>