generic/711,xfs/537: actually fork these tests for exchange-range
Fork g/711 to g/752, x/537 to x/615. These tests to check the same
things with exchange-range as they do for swapext, since the code
porting swapext to commit-range has been dropped.

I was going to fork xfs/789 as well, but it turns out that generic/714
covers this sufficiently so for that one, we just strike fiexchange
from the group tag.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Zorro Lang <zlang@kernel.org>