remove spurious _supported_fs calls
Remove _supported_fs calls for generic in the generic directory
or for $FSTYP in the $FSTYP directory.

This leaves us with the negative checks, and the overloaded ext4
directory where some tests can also be run for ext2 and ext3.

While at it also remove the pointless "real QA test starts here"
usually placed right next to it.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: David Disseldorp <ddiss@suse.de>
Reviewed-by: Darrick J. Wong <djwong@kernel.org>
Signed-off-by: Zorro Lang <zlang@kernel.org>