generic/019: redirect fsstress output to log file instead
Currently we are redirecting stdout and stderr of fsstress to /dev/null.
In case we have a test failure, it's useful to know the seed that
fsstress was using because it might be helpful to run it again with that
same seed in the hope of being able to reproduce the failure.

So redirect stdout/stderr to the log file $seqres.full instead, so
that we'll see a line like this after running the test:

    seed = 1722389488

Signed-off-by: Filipe Manana <fdmanana@suse.com>
Reviewed-by: David Sterba <dsterba@suse.com>
Reviewed-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: Zorro Lang <zlang@kernel.org>