xfs/229: call on the test directory
xfs/229 operates on a directory that is forced to the data volume, but
it calls _require_fs_space on $TEST_DIR which might point to the RT
device when -d rtinherit is set.

Call _require_fs_space on $TDIR after it is created to check for the
space actually used by the test.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Darrick J. Wong <djwong@kernel.org>
Reviewed-by: "Hans Holmberg" <hans.holmberg@wdc.com>
Signed-off-by: Zorro Lang <zlang@kernel.org>